Added exception handling to boost score to 60/80
Problem with binary file storage/retrieval.  Headers?

Mpeg video data can be submitted for a video
Rubric Score: 0.0/20.0
Goal: The goal of this evaluation is to ensure that your Spring
controller(s)
allow mpeg video data to be submitted for a video. The test also checks
that the
controller(s) can serve that video data to the client.