Submitted Assignment
Assignment Evaluation
Your solution received points for the following grading rubrics:
Video data is preserved
Rubric Score: 20.0/20.0
Goal: The goal of this evaluation is to ensure that your Spring
controller(s) properly unmarshall Video objects from the data that is
sent to them and that the HTTP API for adding videos is implemented
properly. The test checks that your code properly accepts a request
body with application/json data and preserves all the properties that
are set by the client. The test also checks that you generate an ID and
data URL for the uploaded video.
Reference: This test is derived from the material in these videos:
https://class.coursera.org/mobilecloud-001/lecture/61
https://class.coursera.org/mobilecloud-001/lecture/67
https://class.coursera.org/mobilecloud-001/lecture/71

The list of videos is updated after an add
Rubric Score: 20.0/20.0
Goal: The goal of this evaluation is to ensure that your Spring
controller(s) can add videos to the list that is stored in memory on
the server. The test also ensure that you properly return a list of
videos as JSON.
Reference: This test is derived from the material in these videos:
https://class.coursera.org/mobilecloud-001/lecture/61
https://class.coursera.org/mobilecloud-001/lecture/67
https://class.coursera.org/mobilecloud-001/lecture/71

Mpeg video data can be submitted for a video
Rubric Score: 20.0/20.0
Goal: The goal of this evaluation is to ensure that your Spring
controller(s) allow mpeg video data to be submitted for a video. The
test also checks that the controller(s) can serve that video data to
the client.
Reference: This test is derived from the material in these videos:
https://class.coursera.org/mobilecloud-001/lecture/69
https://class.coursera.org/mobilecloud-001/lecture/65
https://class.coursera.org/mobilecloud-001/lecture/71
https://class.coursera.org/mobilecloud-001/lecture/207

Requests for non-existant video data return a 404
Rubric Score: 10.0/10.0
Goal: The goal of this evaluation is to ensure that your Spring
controller(s) properly indicate to the client with a 404 response when
the client sends a request for video data for a video that does not
have any video data.
Reference: This test is derived from the material in these videos:
https://class.coursera.org/mobilecloud-001/lecture/65
https://class.coursera.org/mobilecloud-001/lecture/207

Attempting to submit video data for a non-existant video generates a 404
Rubric Score: 10.0/10.0
Goal: The goal of this evaluation is to ensure that your Spring
controller(s) produce a 404 error if a client attempts to submit video
data for a video that does not exist.
Reference: This test is derived from the material in these videos:
https://class.coursera.org/mobilecloud-001/lecture/207
https://class.coursera.org/mobilecloud-001/lecture/69
https://class.coursera.org/mobilecloud-001/lecture/65